+++ tgbot/filters/car_in_db.py
@@ -19,7 +19,7 @@ class CarInDB(BoundFilter):
         session_maker = obj.bot.get('db')
 
         car_number: types.Message.text = obj.text
-        car = await Car.get_active_car(session_maker, car_number)
+        car = await Car.get_active_car(session_maker, car_number.upper())
 
         data = dict(car=car)
         if car is None:
