+++ tgbot/handlers/user_menu.py
@@ -28,12 +28,12 @@ async def feedback(call: CallbackQuery):
 
 
 async def about_us(call: CallbackQuery):
-    await call.message.edit_text(f"𝐖𝐈𝐔𝐓 𝐏𝐚𝐫𝐤𝐢𝐧𝐠 𝐛𝐨𝐭\n"
+    await call.message.edit_text(f"𝐖𝐈𝐔𝐓 𝐏𝐚𝐫𝐤𝐢𝐧𝐠 𝐛𝐨𝐭\n\n"
                                  f"This bot exists thankfully for those who contributed\n"
                                  f"this <a href='https://github.com/mad-maids/maid.ritsu'>project</a>"
                                  f", and they are:\n\n"
-                                 f"👨‍💻<a href='tg://user?id=698728556'>Azizbek</a> (Co-Creator, Maintainer)\n"
-                                 f"🕵️‍♂<a href='https://t.me/muminovbob'>Bobomurod</a> (Co-Creator, Maintainer)\n"
+                                 f"👨‍💻<a href='tg://user?id=5560163548'>Azizbek</a> (Co-Creator, Maintainer)\n"
+                                 f"🕵️‍♂<a href='tg://user?id=126073578'>Bobomurod</a> (Co-Creator, Maintainer)\n"
                                  f"👩‍🚀<a href='https://github.com/uwussimo'>UwUssimo</a> (Core Contributor)\n\n"
                                  f"Copyright © 2022 <a href='https://github.com/mad-maids'>Mad Maids</a>",
                                  reply_markup=about_us_keyboard,
@@ -87,7 +87,8 @@ async def helper(msg: Message):
     await msg.answer(help_text)
 
 
-async def reporter(msg: Message):
+async def reporter(msg: Message, state: FSMContext):
+    await state.finish()
     text = "<code>If you have been faced on of the troubles below, \n" \
            "please, choose related one and admins will do their best to solve it!</code>"
     await msg.answer(text, reply_markup=report_keyboad)
